requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
Django==1.10.1
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-js-asset==1.0.0
idna==2.6
lxml==4.2.1
MySQL-python==1.2.5
Pillow==5.1.0
python-docx==0.8.6
requests==2.18.4
urllib3==1.22
